+++ xenon-common/src/main/java/com/vmware/xenon/services/common/QueryPageService.java
@@ -43,6 +43,12 @@ public class QueryPageService extends StatelessService {
     public static class LuceneQueryPage {
         public String previousPageLink;
         public ScoreDoc after;
+        public Integer groupOffset;
+
+        public LuceneQueryPage(String link, int groupOffset) {
+            this.previousPageLink = link;
+            this.groupOffset = groupOffset;
+        }
 
         public LuceneQueryPage(String link, ScoreDoc after) {
             this.previousPageLink = link;